Gittle: drop superfluous stuff
Since the head property returns an SHA, and the active_sha
function returns an SHA for head (after apparently spinning its
wheels for a bit), I'm reasonably sure that the active_sha
property is in fact equivalent to the head property.

If so, it should probably be dropped (as its presence suggests
that the two properties do something different).  Taking this
belief to be correct as a given, I've clarified it by replacing
the old implementation with a new one: "return head", documented
that the API is deprecated, and noted why.

This leaves zero consumers of _active_branch, hence, I've
eliminated it.

Signed-off-by: Gregory M. Turner <gmt@be-evil.net>